src/main/java/racingcar/domain/CarFactory.java
package racingcar.domain;

import java.util.List;

public class CarFactory {

    private static CarFactory carFactory;

    private CarFactory() {
    }

    public static CarFactory getInstance() {
        if (carFactory == null) {
            carFactory = new CarFactory();
        }
        return carFactory;
    }

    public Cars createCars(List<String> carNames) {
        List<Car> carList = carNames.stream()
                .map(Car::new)
                .toList();

        return new Cars(carList);
    }
}
